Refactoring done

+++ Framework/Include/Scene/Scene.h
@@ -37,7 +37,7 @@ namespace Trinity
 			glm::vec3 position;
 			glm::vec3 normal;
 			glm::vec2 uv;
-			glm::vec4 joint;
+			glm::uvec4 joint;
 			glm::vec4 weight;
 		};
